+++ utils/misc_test.go
@@ -20,6 +20,7 @@
 package utils
 
 import (
+	"github.com/apache/incubator-openwhisk-wskdeploy/dependencies"
 	"github.com/stretchr/testify/assert"
 	"os"
 	"testing"
@@ -44,29 +45,17 @@ func TestURLReader_ReadUrl(t *testing.T) {
 }
 
 func TestDependencies(t *testing.T) {
-	var record = NewDependencyRecord("projectPath", "packageName", "http://github.com/user/repo", "master", nil, nil, false)
+	var record = dependencies.NewDependencyRecord("projectPath", "packageName", "http://github.com/user/repo", "master", nil, nil, false)
 	assert.Equal(t, "projectPath", record.ProjectPath, "ProjectPath is wrong")
 	assert.Equal(t, "http://github.com/user/repo", record.Location, "URL is wrong")
 	assert.Equal(t, "http://github.com/user/repo", record.BaseRepo, "BaseRepo is wrong")
 	assert.Equal(t, "", record.SubFolder, "SubFolder is wrong")
 
-	record = NewDependencyRecord("projectPath", "packageName", "http://github.com/user/repo/subfolder1/subfolder2", "master", nil, nil, false)
+	record = dependencies.NewDependencyRecord("projectPath", "packageName", "http://github.com/user/repo/subfolder1/subfolder2", "master", nil, nil, false)
 	assert.Equal(t, "projectPath", record.ProjectPath, "ProjectPath is wrong")
 	assert.Equal(t, "http://github.com/user/repo/subfolder1/subfolder2", record.Location, "URL is wrong")
 	assert.Equal(t, "http://github.com/user/repo", record.BaseRepo, "BaseRepo is wrong")
 	assert.Equal(t, "/subfolder1/subfolder2", record.SubFolder, "SubFolder is wrong")
-}
-
-func TestParseOpenWhisk(t *testing.T) {
-	openwhiskHost := "https://openwhisk.ng.bluemix.net"
-	openwhisk, err := ParseOpenWhisk(openwhiskHost)
-	assert.Equal(t, nil, err, "parse openwhisk info error happened.")
-	converted := ConvertToMap(openwhisk)
-	assert.Equal(t, 2, len(converted["nodejs"]), "not expected length")
-	assert.Equal(t, 1, len(converted["php"]), "not expected length")
-	assert.Equal(t, 1, len(converted["java"]), "not expected length")
-	assert.Equal(t, 3, len(converted["python"]), "not expected length")
-	assert.Equal(t, 1, len(converted["swift"]), "not expected length")
 }
 
 func TestNewZipWritter(t *testing.T) {
